+++ seleniumwire/__init__.py
@@ -3,4 +3,4 @@
 """Top-level package for Selenium Wire."""
 
 __author__ = """Will Keeling"""
-__version__ = '1.0.9'
+__version__ = '1.0.10'