+++ HISTORY.rst
@@ -1,6 +1,12 @@
 History
 ~~~~~~~
 
+1.0.10 (2019-09-22)
+------------------
+
+* Scope request capture.
+* Apply header filtering on a per-URL basis.
+
 1.0.9 (2019-08-25)
 ------------------